@@ -13,87 +12,55 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "^4.1.0",
-    "@angular/compiler": "4.1.0",
-    "@angular/compiler-cli": "4.1.0",
-    "@angular/core": "4.1.0",
-    "@angular/forms": "4.1.0",
-    "@angular/http": "4.1.0",
-    "@angular/platform-browser": "^4.1.0",
-    "@angular/platform-browser-dynamic": "4.1.0",
-    "@ionic-native/app-version": "^3.10.2",
-    "@ionic-native/core": "3.7.0",
-    "@ionic-native/device": "^3.10.2",
-    "@ionic-native/facebook": "^3.10.3",
-    "@ionic-native/file": "^3.10.3",
-    "@ionic-native/geolocation": "^3.10.2",
-    "@ionic-native/keyboard": "^3.10.3",
-    "@ionic-native/onesignal": "^3.12.1",
-    "@ionic-native/splash-screen": "3.7.0",
-    "@ionic-native/status-bar": "3.7.0",
-    "@ionic-native/transfer": "^3.10.3",
-    "@ionic/storage": "2.0.1",
-    "cordova-android": "6.4.0",
-    "cordova-ios": "^4.5.2",
+    "@angular/common": "5.0.0",
+    "@angular/compiler": "5.0.0",
+    "@angular/compiler-cli": "5.0.0",
+    "@angular/core": "5.0.0",
+    "@angular/forms": "5.0.0",
+    "@angular/http": "5.0.0",
+    "@angular/platform-browser": "5.0.0",
+    "@angular/platform-browser-dynamic": "5.0.0",
+    "@ionic-native/app-version": "^4.4.0",
+    "@ionic-native/core": "4.3.3",
+    "@ionic-native/device": "^4.4.0",
+    "@ionic-native/file": "^4.4.0",
+    "@ionic-native/file-transfer": "^4.4.0",
+    "@ionic-native/geolocation": "^4.4.0",
+    "@ionic-native/native-storage": "^4.4.0",
+    "@ionic-native/splash-screen": "4.3.3",
+    "@ionic-native/status-bar": "4.3.3",
+    "@ionic/pro": "1.0.14",
+    "@ionic/storage": "^2.1.3",
     "cordova-plugin-app-version": "^0.1.9",
-    "cordova-plugin-compat": "^1.2.0",
-    "cordova-plugin-console": "^1.1.0",
-    "cordova-plugin-device": "^1.1.6",
-    "cordova-plugin-facebook4": "^1.9.1",
-    "cordova-plugin-file": "^4.3.3",
-    "cordova-plugin-file-transfer": "^1.6.3",
-    "cordova-plugin-geolocation": "^2.4.3",
-    "cordova-plugin-googlemaps": "^2.1.1",
-    "cordova-plugin-googlemaps-sdk": "git+https://github.com/mapsplugin/cordova-plugin-googlemaps-sdk.git",
+    "cordova-plugin-device": "^1.1.4",
+    "cordova-plugin-file": "^5.0.0",
+    "cordova-plugin-file-transfer": "^1.7.0",
+    "cordova-plugin-geolocation": "^3.0.0",
     "cordova-plugin-nativestorage": "^2.2.2",
-    "cordova-plugin-splashscreen": "^4.0.3",
-    "cordova-plugin-statusbar": "^2.2.3",
-    "cordova-plugin-whitelist": "^1.3.2",
-    "cordova-uxcam": "^2.3.6",
-    "intl": "^1.2.5",
-    "ionic-angular": "3.2.1",
-    "ionic-image-loader": "^3.2.0",
-    "ionic-plugin-keyboard": "^2.2.1",
+    "ionic-angular": "3.9.2",
+    "ionic-image-loader": "^4.2.1",
     "ionicons": "3.0.0",
-    "onesignal-cordova-plugin": "^2.2.2",
-    "rxjs": "5.1.1",
+    "rxjs": "5.5.2",
     "sw-toolbox": "3.6.0",
-    "zone.js": "0.8.10"
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "1.3.7",
+    "@ionic/app-scripts": "3.1.0",
     "compass": "^0.1.1",
     "compass-sass-mixins": "^0.12.7",
-    "ionic": "3.18.0",
-    "typescript": "2.2.1"
+    "typescript": "2.4.2"
   },
+  "description": "An Ionic project",
   "cordova": {
     "plugins": {
       "cordova-plugin-nativestorage": {},
       "cordova-plugin-device": {},
-      "cordova-plugin-console": {},
-      "cordova-plugin-whitelist": {},
-      "ionic-plugin-keyboard": {},
       "cordova-plugin-app-version": {},
-      "cordova-uxcam": {},
-      "cordova-plugin-geolocation": {},
+      "cordova-plugin-geolocation": {
+        "GEOLOCATION_USAGE_DESCRIPTION": "Identificar o seu local para entregar a cerveja"
+      },
       "cordova-plugin-file": {},
-      "cordova-plugin-file-transfer": {},
-      "cordova-plugin-facebook4": {
-        "APP_ID": "289767478086559",
-        "APP_NAME": "Cerveja.me"
-      },
-      "cordova-plugin-splashscreen": {},
-      "cordova-plugin-statusbar": {},
-      "cordova-plugin-googlemaps": {
-        "API_KEY_FOR_ANDROID": "AIzaSyAKiUVVFdR0OlfrR34QtKu9sBf-oyYuiGs",
-        "API_KEY_FOR_IOS": "AIzaSyAKiUVVFdR0OlfrR34QtKu9sBf-oyYuiGs"
-      },
-      "onesignal-cordova-plugin": {}
-    },
-    "platforms": [
-      "android",
-      "ios"
-    ]
+      "cordova-plugin-file-transfer": {}
+    }
   }
 }
