@@ -1,8 +1,7 @@
-import { Component } from '@angular/core';
-import { Platform } from 'ionic-angular';
+import { Component, ViewChild } from '@angular/core';
+import { Nav, Platform } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
-
 import { Storage } from '@ionic/storage';
 
 import { HomePage } from '../pages/home/home';
@@ -12,36 +11,43 @@ import { TourPage } from '../pages/tour/tour';
   templateUrl: 'app.html'
 })
 export class MyApp {
-  rootPage:any;
+  // @ViewChild(Nav) nav: Nav;
+
+  rootPage: any; // = HomePage;
+
+  // pages: Array<{title: string, component: any}>;
 
   constructor(
     public platform: Platform,
     public statusBar: StatusBar,
     public splashScreen: SplashScreen,
-    public storage: Storage
+    private storage:Storage) {
+      this.storage.get('hasSeenTutorial')
+      .then((hasSeenTutorial) => {
+        if (hasSeenTutorial) {
+          this.rootPage = HomePage;
+        } else {
+          this.rootPage = TourPage;
+        }
 
-    ) {
-    this.storage.get('hasSeenTutorial')
-    .then((hasSeenTutorial) => {
-      if (hasSeenTutorial) {
-        this.rootPage = HomePage;
-      } else {
-        this.rootPage = TourPage;
-      }
-      this.startApp();
-    });
+        this.initializeApp();
+      })
 
-  }
-  startApp(){
 
-    this.platform.ready().then(() => {
-      // Okay, so the platform is ready and our plugins are available.
-      // Here you can do any higher level native things you might need.
-      this.statusBar.styleDefault();
-      setTimeout(() => {
+    }
+
+    initializeApp() {
+      this.platform.ready().then(() => {
+        // Okay, so the platform is ready and our plugins are available.
+        // Here you can do any higher level native things you might need.
+        this.statusBar.styleDefault();
         this.splashScreen.hide();
-      }, 500);
-    });
-  }
-}
+      });
+    }
 
+    // openPage(page) {
+    //   // Reset the content nav to have just this page
+    //   // we wouldn't want the back button to show in this scenario
+    //   this.nav.setRoot(page.component);
+    // }
+  }
