@@ -2,10 +2,10 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 @Injectable()
-export class VoucherProvider {
+export class HttpProvider {
 
   constructor(public http: HttpClient) {
-
+    console.log('Hello HttpProvider Provider');
   }
 
 }
