@@ -1,5 +1,5 @@
 /**
- * Check out https://googlechrome.github.io/sw-toolbox/ for
+ * Check out https://googlechromelabs.github.io/sw-toolbox/ for
  * more info on how to use sw-toolbox to custom configure your service worker.
  */
 
@@ -15,6 +15,7 @@ self.toolbox.options.cache = {
 self.toolbox.precache(
   [
     './build/main.js',
+    './build/vendor.js',
     './build/main.css',
     './build/polyfills.js',
     'index.html',
@@ -23,7 +24,7 @@ self.toolbox.precache(
 );
 
 // dynamically cache any other local assets
-self.toolbox.router.any('/*', self.toolbox.cacheFirst);
+self.toolbox.router.any('/*', self.toolbox.fastest);
 
 // for any other requests go to the network, cache,
 // and then only use that cached resource if your user goes offline
